Completely reworked the Tabs component
This started as a proof of concept, but I restructured the whole
component with the inital goal of adding support for tab changes without
the content elements  https://github.com/vuetifyjs/vuetify/issues/273.
The result still needs some work, but it is functioning. Here is a list
of the major changes.
1. Created a silder component that takes the current tab as a prop and
then adjusts accordingly.
2. Removed the need for slot defintions when creating the components
(render function takes care of spliiting based on tag-name)
3. Used the render function to also filter out tab-content if it didn't
have a matching id
4. Removed a bunch of code and imports that are no longer necessary
5. Hoefully created a leaner component

I've commented in where some parts may still need some tweaking or could
be improved. The only thing not working from the original is the forward
animation.

I realize that this diverges pretty drastically from the original, but I
thought I would commit this for at least your consideration.