Fix/2146 (#2150)
* fix($browser): iPhone iOS click-outside bug

iOS does not allow click events to be bound to the body, this is the entire purpose of v-app. Was
mistakenly allowed to be removed with PR #1932

* fix($browser): Edge bug with for..of

The new z-index functionality utilizes for..of quite extensively. When a HTMLElementList is iterated
on using for..of and the list is empty, causes Edge to throw a Symbol.iterator fatal

fixes #2146

* refactor: Change forEach to an iterator & update comments to be clearer

Converting the NodeList to an array is enough to fix the issue, but this
wasn't reflected in the comments

* Update click-outside.js