fix(v-navigation-drawer): improved model interaction and functionality
updated logic to always prioritize a provided model. if undefined or null is provided, the drawer
will determine if to open based upon the props used, **permanent** or **temporary** or nothing.
changed **enable-resize-watcher** to implicit with a new **disable-resize-watcher** prop. added
better support when switching between different `v-navigation-drawer` types.

fixes #2314