feat(v-list): added nesting functionality to list-groups (#2563)
* feat(v-list): added nesting functionality to list-groups

added registrable to list components. added ability to nest groups with new **sub-group** prop.
added default icons for `v-list-group` append and prepend icon with optional slot to override. added
group highlighting similar to google analytics

* feat(text): added text helper classes from spec

this resolves the issue with list tile active and regular colors. can now have a specific color and
a different one for when the router-link is active

resolves #1898

* removed commented code and fixed slot

* updated snapshots, updated registrable

* added more unit tests

* fixes codebeat

* fixed disabled states