Fix ripple directive binding issues
Unbind was incorrect, it did not remove the registered event handlers.
Update did not update the value used by the event handler,
so a checkbox that is initially disabled would have an event
handler without the center: true option, and enabling the
checkbox would cause the ripple to appear at the mouse pointer's
location.

Fixes #2677