fix: Avoid high CPU usage in Safari
A v-navigation-drawer with the app param causes Safari to repeatedly
invalidate and recalculate the layout, permanently consuming a CPU
core.

This is a bug in Safari (fixed in current tech preview) which is
triggered by the combination of using calc in the max-height,
transition-property including max-height and the element being created
using document.createElement (as vue does).

fixes #2773