fix: `update:returnValue` event emitted twice in returnable mixin
The `returnable` mixin emits `update:returnValue` event twice on save.
The first time inside `save` method, the second time directly after
this when the value of `isActive` flag is changed to `false`
(in this case the watcher emits the event), so it is enough to just
set the `isActive` flag to false inside `save` method and the
`update:returnValue` event will be called "automatically" by watcher.