Fix #5843 v-menu hover stuck open
Fix for https://github.com/vuetifyjs/vuetify/issues/5843 - in the event isActive has been reset to false between the time runDelay(open) was called and when we were provisioned this animation frame, sync up hasJustFocused to false and do a no-op assign to isContentActive which doesn't trigger the watcher. Resolve the promise although it's not really needed, just so we aren't breaking promises.