[fix] Make VTimePicker select hours after having selected minutes
After selecting minutes (or seconds if enabled), the TimePicker stayed
in minutes (seconds) selection.
When used inside a VMenu the TimePicker
would open in minute-mode when the Menu is re-opened unless destroyed
on close.
When used with the `no-title` prop there is not even a way to switch
back to hour selection

See vuetifyjs/vuetify#1549, fixes vuetifyjs/vuetify#4502