test(VSlider): Check roundValue is bounded by min and max
The roundValue function gets used by inputWidth to determine placement of the thumb and various
other elements. If the initial value is less than min, inputWidth will return a negative number due
to the roundValue not being bounded by min and max. This test will check the boundaries.

test for fix #6604