feat(VTreeview): add selectionType prop (#7575)
* fix(VTreeview): unable to unselect child nodes in certain scenarios

this is a breaking change. parent nodes are no longer included in the selection.
instead two different selection types are offered, leaf and independent. leaf is
similar to existing behaviour except as mentioned parents are not included.
independent selects nodes independent of their position in tree hierarchy, and
allows for selection of parent nodes. selection does not propagate up or down.

solves #6903

* docs(VTreeview): updated docs with selection-type prop info and example

* fix(VTreeview): updated tests

* chore: renamed example file

* Update packages/vuetify/src/components/VTreeview/VTreeview.ts

Co-Authored-By: Dmitry Sharshakov <d3dx12.xx@gmail.com>