fix(overlayable): #8625
Fixes overlayable keeping VOverlay open on background tabs.

Most modern browsers restrict requestAnimationFrame when the tab is inactive. My understanding is that since the requestAnimationFrame is delayed, the when the overlay attempts to remove itself with the event listener the check to destroy the overlay returns from the event listener because the requestAnimationFrame sets this.overlay.value to true.