feat(VIcon): add support for FA5 SVG icons (#9995)
Font Awesome icons does not returns directly an SVG path like MDI, but an object containing more informations.
One very useful information is the viewbox size, which was defaulting to 24 within VIcon.
This commit brings a new interface VuetifyIconSVG which can be given directly as the slot of VIcon, to handle those cases.
It also add the support for a Font Awesome SVG icon definition object to be passed directly within the VIcon slot (it will internaly be converted to a VuetifyIconSVG).