fix(VDataTable): union column filter with search results (#11575)
fixes #11402

* fix: 🐛 Fix behaviour when combining filters and search terms

Ensure that filters and search terms are combined to display the
intersection of results, rather than the union - so each filter applied
makes the set of matches smaller for refined searching

* fix: 🐛 Fix lint error

Sorry, seems to have run locally with a different setting! Will look
into that.. :(

* refactor: 💡 Fixing another liniting issue

Local setup appears to be running with different rules - sorry!