fix: revert removal of ripple's blocking of title hover (#12775)
The `pointer-events: none` workaround had caused a new issue of
ripple not disappearing on mouseout event and has been removed.

As we've moved the title attribute to a parent v-input, it's no
longer necessary to disable the ripple's hover effect override,
because ripple now properly inherits the title. We'd still need
some new workaround if we had to explicitly put title on input.