fix(voverlay): show overlay with :model-value=true in correct location
Menus and tooltips with :model-value='true' were not in the correct position because the refs for
attached element were still undefined when the watchEffect in useLocationStrategies finished
running. This fix will cause the watchEffect to run a second time after render, but only when the
overlay is attached to an element.

fix #16042