fix(navigation-drawer): fix active state not updated on resize with rail
When having rail as a prop and v-model i noticed that while resizing the drawer scrim stayed on top
and application became unusable because isActive value stayed benig true. Thats why I added nextTick
on the update of isActive so when all resizing is done and DOM is updated, we update isActive.

fix #15807