docs(upgrade-guide): add some renamed components (#16635)
## Description

Add missing obsolete or renamed components.

I have a discussion about it on Vuetify Discord #documentation 

> Hello, I will have to migrate a large application from vue 2 to vue 3
(using vuetify). I am making an inventory of the components used, and if
they are still present and how to migrate them. There are some
components, that I didn't find in vuetify 3: v-simple-table,
v-subheader, v-tab-item. What are the status about them ? Should we add
information about this in the migration guide ?

> johnleider — Today à 16:00
> v-simple-table -> v-table
> v-subheader -> v-list-subheader
> v-tab-item -> v-window-item