Doc(tables): Make clear that v-data-table exists
I wanted a Vue3 framework with good table support, and looked at Vuetify. I saw a heading "Data and display" and a single entry "Tables". When I clicked that, I came to https://vuetifyjs.com/en/components/tables/, which implied that there was _only_ the `v-table` component. I almost abandoned Vuetify, but by chance googled for "Vuetify Sortable Table" looking for a 3rd party component, and via that discovered the `vue-data-table` component. 

It is vital that _either_ both tables are listed in the left-hand menu of "components", or the `v-table` entry clearly links to the `v-data-table` component very early in the text. Currently the only link is at the bottom, but entitled "Basics" which I did not see any reason to click, because I thought it would be _even more basic_ than v-table, 8-)