fix(VImg): clear polling timer on component unmount
Fixes an issue where a component can be destroyed before the timer
completes, causing the callback to be called on an invalid object. Also
moves from window.setTimeout to global setTimeout, which seems to be the
standard in other files. Calling the callback of a destoryed component,
combined with the use of window.setTimeout, caused tests run with vitest
to throw random unhandledexceptions because window was no longer
defined. Without having the timer cleared in unmount, there's no good
way to ensure it will not trigger after tests clean up.

fixes #18072