Add types prop back to v-skeleton-loader
I merge the custom `types` prop into the `rootTypes` and set that as a computed to keep it reactive.

I opted to move the gen* functions into setup to inherit the `mergedTypes` computed rather than pass it around as an extra parameter. Happy to set it up as a parameter instead if desired.