fix: type compatibility with vue 3.4
This PR updates the vue version to 3.4.4 and fixes the type
compatibility issues.

Vue 3.4 introduced a type-only breaking change that requires the project
to explicitly import `vue/jsx` to register the global JSX namespace.
https://blog.vuejs.org/posts/vue-3-4#global-jsx-namespace

`vue/jsx` is available since Vue 3.3
https://blog.vuejs.org/posts/vue-3-3#jsx-import-source-support
So the peer dependency version range is not changed in this PR,
therefore not a breaking change to the project.