fix(VMenu): allow enter keypress to work in textareas
Previously, when a `<v-menu />` with `:close-on-content-click="false"`
contained a textarea, an enter press inside the textarea would do
nothing and not produce a new line. Now, it will allow a new line to be
typed into the textarea when pressing enter.

This last worked in Vuetify v3.5.14 and was broken starting in v3.5.15.
The fix for #19519 in 85ba4a800 caused this issue.

fixes #19767