feat(docs): add Figma link support to examples
Add support for displaying Figma design links in example
components. Examples can now include an example-meta block
with a figma URL that will be displayed as an action button
in the example toolbar.

The exampleMeta computed property was designed with
extensibility in mind, allowing for future integrations
with other tools such as Storybook.

- Add exampleMeta computed property to parse meta data
- Add figma icon alias to vuetify plugin
- Add view-in-figma i18n translation
- Add example-meta block support in vite config
- Add example-meta to v-btn readonly example